Add bucket lists for (upcoming) Hector events
Ongoing or past events that have teams (pairs) no longer display
the bucket lists because we'd need the participants handicaps at
the time and we don't have that information at hand.

If we get those historical handicaps, we can allow past events
display the buckets as well, pulling the handicaps from the
handicap history rather than the player object.